[DEV-3101] fix circular dependency issue
