+++ src/subdomains/core/payment-link/dto/payment-link.dto.ts
@@ -1,10 +1,9 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
-import { Type } from 'class-transformer';
-import { IsEmail, IsOptional, IsString, IsUrl, ValidateNested } from 'class-validator';
 import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
 import { ErrorDto } from 'src/shared/dto/error.dto';
 import { PaymentLinkPaymentMode, PaymentLinkPaymentStatus, PaymentLinkStatus, PaymentStandard } from '../enums';
 import { PaymentLinkConfigDto } from './payment-link-config.dto';
+import { PaymentLinkRecipientDto } from './payment-link-recipient.dto';
 
 export type TransferMethod = Blockchain;
 
@@ -62,61 +61,6 @@ export interface PaymentLinkEvmPaymentDto {
 
 export interface PaymentLinkHexResultDto {
   txId: string;
-}
-
-export class PaymentLinkRecipientAddressDto {
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  street?: string;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  houseNumber?: string;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  zip?: string;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  city?: string;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  country?: string;
-}
-
-export class PaymentLinkRecipientDto {
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  name?: string;
-
-  @ApiPropertyOptional({ type: PaymentLinkRecipientAddressDto })
-  @IsOptional()
-  @Type()
-  @ValidateNested()
-  address?: PaymentLinkRecipientAddressDto;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  phone?: string;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsEmail()
-  mail?: string;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsUrl()
-  website?: string;
 }
 
 export class PaymentLinkPaymentDto {
