+++ src/subdomains/core/payment-link/entities/payment-link.config.ts
@@ -1,5 +1,5 @@
 import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
-import { PaymentLinkRecipientDto } from '../dto/payment-link.dto';
+import { PaymentLinkRecipientDto } from '../dto/payment-link-recipient.dto';
 import { PaymentQuoteStatus, PaymentStandard } from '../enums';
 
 export interface PaymentLinkConfig {
